requirements.txt
Flask==2.1.1
google_api_python_client==2.42.0
influxdb==5.3.1
matplotlib==3.5.1
numpy==1.20.1
paho_mqtt==1.6.1
pandas==1.4.2
protobuf==3.20.0
sense_hat==2.2.0
